Db UPdated

+++ src/IICT-Store.Models/Products/Distribution.cs
@@ -33,7 +33,7 @@ namespace IICT_Store.Models.Products
         public long? ReceiverId { get; set; }
         public Person Receiver { get; set; }
 /*        public ICollection<ProductSerialNo> ProductSerialNo { get; set; }*/
-        public int DistributedTo { get; set; } // to samir sir/ room no 212
+        public long DistributedTo { get; set; } // to samir sir/ room no 212
         public string SignatureOfReceiver { get; set; }
         public string Note { get; set; }
     }
